[main] Ajustar: decimal flotante. Ajustar templates: exlcuir propiedades borradas, unificar archivos templates. Agregar: visualizar templates.

+++ namasdev.Apps/namasdev.Apps.Entidades/PropiedadTipoEspecificacionesDecimal.cs
@@ -1,6 +1,6 @@
 ﻿namespace namasdev.Apps.Entidades
 {
-    public class PropiedadTipoEspecificacionesDecimal : PropiedadTipoEspecificacionesDecimalBase<double>
+    public class PropiedadTipoEspecificacionesDecimal : PropiedadTipoEspecificacionesDecimalBase<decimal>
     {
     }
 }
